+++ service/controller/azuremachinepool/handler/nodepool/template/template.go
@@ -2,6 +2,7 @@ package template
 
 import (
 	"encoding/json"
+	"fmt"
 	"reflect"
 
 	azureresource "github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2019-05-01/resources"
@@ -63,6 +64,7 @@ func Diff(currentDeployment azureresource.DeploymentExtended, desiredDeployment 
 	if currentParameters.ClusterID != desiredParameters.ClusterID {
 		changes = append(changes, "clusterID")
 	}
+	fmt.Printf("current = %q, desired = %q\n", currentParameters.KubernetesVersion, desiredParameters.KubernetesVersion)
 	if currentParameters.KubernetesVersion != desiredParameters.KubernetesVersion {
 		changes = append(changes, "kubernetesVersion")
 	}