+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.7.1] - 2021-06-21
+
 ### Changed
 
 - Consider node pools out of date if flatcar image has changed.
@@ -387,7 +389,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/azure-operator/compare/v5.7.0...HEAD
+[Unreleased]: https://github.com/giantswarm/azure-operator/compare/v5.7.1...HEAD
+[5.7.1]: https://github.com/giantswarm/azure-operator/compare/v5.7.0...v5.7.1
 [5.7.0]: https://github.com/giantswarm/azure-operator/compare/v5.6.0...v5.7.0
 [5.6.0]: https://github.com/giantswarm/azure-operator/compare/v5.5.2...v5.6.0
 [5.5.2]: https://github.com/giantswarm/azure-operator/compare/v5.5.1...v5.5.2